[Polly] [WIP] [Invariant Loads] Teach PPCGCodeGeneration to use ivnariant loads correctly.
We need to relax constraints on invariant loads so that they do not
create fake RAW dependences. So, we do not consider invariant loads as
scalar deps in a region.

Also, we teach PPCGCodeGeneration to consider the ValueMap when
generating code that uses Values, so it picks up the correct IR value
(which has been load hoisted).