[Polly] Model multi-dimensional memory accesses as function calls. [WIP]
Summary:
This is a proof-of-concept patch that shows how a function call can model
a multi-dimensional memory access. It can be used to model sparse matrices or
other multi-dimensional accesses which we cannot easily delinearize.

I currently do not intend to upstream this change, but wanted to push it out as
this came up in a recent discussion.

Reviewers: pollydev

Tags: #polly

Differential Revision: https://reviews.llvm.org/D29583